requirements.txt
Akhet==1.0.2
Beaker==1.6.2
Chameleon==2.7.3
Mako==0.6.2
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.5.0
Pygments==1.4
SQLAHelper==1.0
SQLAlchemy==0.7.5
Tempita==0.5.1
WebError==0.10.3
WebHelpers==1.3
WebOb==1.2b3
amqplib==1.0.2
anyjson==0.3.1
chatty==0.0
demjson==1.6
distribute==0.6.24
kombu==2.1.0
morbid==0.8.7.3
orbited==0.7.10
psycopg2==2.4.4
pyramid==1.3a7
pyramid-beaker==0.6.1
pyramid-debugtoolbar==0.9.8
pyramid-mailer==0.6
pyramid-tm==0.3
repoze.lru==0.4
repoze.sendmail==2.3
simplejson==2.3.2
stomper==0.2.4
transaction==1.2.0
translationstring==1.1
venusian==1.0a2
waitress==0.8
wsgiref==0.1.2
yolk==0.4.3
zope.deprecation==3.5.0
zope.interface==3.8.0
zope.sqlalchemy==0.7
